Moved functions for estimating guided beam trajectories from StageHipace to Stage.
